requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
boto3==1.28.55
botocore==1.31.85
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.6
MarkupSafe==3.0.2
pika==1.3.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.6.3
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
rococo==1.0.23
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
surrealdb==0.3.2
typing_extensions==4.12.2
urllib3==2.0.7
websockets==10.4
Werkzeug==3.0.4